chore: DTO 하위클래스 작업
