--- src/main/java/com/github/supercodingfinalprojectbackend/dto/request/ChatRoomListRequest.java
@@ -1,9 +1,0 @@
-package com.github.supercodingfinalprojectbackend.dto.request;
-
-
-import lombok.Data;
-
-@Data
-public class ChatRoomListRequest {
-    private Long userId;
-}
