--- src/main/java/com/github/supercodingfinalprojectbackend/dto/request/CreateChatRoomRequest.java
@@ -1,9 +1,0 @@
-package com.github.supercodingfinalprojectbackend.dto.request;
-
-import lombok.Data;
-
-@Data
-public class CreateChatRoomRequest {
-    private Long user1Idx;
-    private Long user2Idx;
-}
