--- src/main/java/com/github/supercodingfinalprojectbackend/dto/response/ResponseChatRoom.java
@@ -1,17 +1,0 @@
-package com.github.supercodingfinalprojectbackend.dto.response;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ResponseChatRoom {
-      private List<ResponseChatRoomInfo> chatRoom;
-}
