--- src/main/java/com/github/supercodingfinalprojectbackend/dto/response/ResponseEnterChatRoom.java
@@ -1,18 +1,0 @@
-package com.github.supercodingfinalprojectbackend.dto.response;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ResponseEnterChatRoom {
-    private List<ResponseChatLog> chatLog;
-}
