+++ src/main/java/com/github/supercodingfinalprojectbackend/repository/ChatRoomRepository.java
@@ -7,9 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
-    List<ChatRoom> findByUser1OrUser2(User user1, User user2);
-    ChatRoom findByChatRoomId(Long chatRoomId);
+    Optional<List<ChatRoom>> findByUser1OrUser2(User user1, User user2);
+    Optional<ChatRoom> findByChatRoomIdAndIsChatIsFalse(Long chatRoomId);
 }
