+++ src/main/java/com/github/supercodingfinalprojectbackend/repository/MessageRepository.java
@@ -2,12 +2,16 @@ package com.github.supercodingfinalprojectbackend.repository;
 
 import com.github.supercodingfinalprojectbackend.entity.ChatRoom;
 import com.github.supercodingfinalprojectbackend.entity.Message;
-import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
+import java.util.Optional;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
-    List<Message> findByChatRoom(ChatRoom chatRoom, Sort sort);
+
+
+    Optional<Page<Message>> findAllByChatRoom(ChatRoom chatRoom, Pageable pageable);
+
     Message findTopByChatRoomOrderBySendAtDesc(ChatRoom chatRoom);
 }
